Created a Guesser to help us guess invalid characters
Started with a method to generate all the variants of a single cell.
Since the spec mentions checking for missing or additional pipes and
underscores, this is based on flipping any of the seven "active" pixels
in a cell (as the top left and top right cells are always spaces).